# docs/swagger.yaml
definitions:
  model.Location:
    properties:
      id:
        type: integer
      id_product:
        type: integer
      name:
        type: string
    type: object
  model.Order:
    properties:
      id:
        type: integer
      id_product:
        type: integer
      id_travel_agent:
        type: integer
      order_address:
        type: string
      order_date:
        type: string
      order_name:
        type: string
      order_phone_number:
        type: string
    type: object
...
  model.ProductPhoto:
    properties:
      file_name:
        type: string
      id:
        type: integer
      id_product:
        type: integer
      photo_name:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: This is Travel Agent System API
  title: Travel Agent System API Documentation
  version: "2.0"
paths:
...
  /location:
    get:
      consumes:
      - application/json
      description: User can get all locations information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
      security:
      - JWT: []
      summary: Get All Locations Information
      tags:
      - Location
    post:
      consumes:
      - application/json
      description: User can create location
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Location'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Location'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Location'
      security:
      - JWT: []
      summary: Create Location
      tags:
      - Location
  /location/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete location information if they want it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Location'
      security:
      - JWT: []
      summary: Delete Location Information
      tags:
      - Location
    get:
      consumes:
      - application/json
      description: User can get location information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Location'
      security:
      - JWT: []
      summary: Get Location Information by Id
      tags:
      - Location
    put:
      consumes:
      - application/json
      description: User can update location information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Location'
      security:
      - JWT: []
      summary: Update Location Information
      tags:
      - Location
...
  /order:
    get:
      consumes:
      - application/json
      description: User can get all orders information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - JWT: []
      summary: Get All Orders Information
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: User can create order
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Order'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - JWT: []
      summary: Create Order
      tags:
      - Order
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete order information if they want it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - JWT: []
      summary: Delete Order Information
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: User can get order information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - JWT: []
      summary: Get Order Information by Id
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: User can update order information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - JWT: []
      summary: Update Order Information
      tags:
      - Order
  /photo:
    get:
      consumes:
      - application/json
      description: User can get all photos information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductPhoto'
      security:
      - JWT: []
      summary: Get All ProductPhotos Information
      tags:
      - ProductPhoto
  /photo/{id}:
    get:
      consumes:
      - application/json
      description: User can get photo information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductPhoto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ProductPhoto'
      security:
      - JWT: []
      summary: Get ProductPhoto Information by Id
      tags:
      - ProductPhoto
...
  /product-photo:
    post:
      consumes:
      - application/json
      description: User can create product photo
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ProductPhoto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProductPhoto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ProductPhoto'
      security:
      - JWT: []
      summary: Create Product Photo
      tags:
      - ProductPhoto
  /product-photo/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete product photo information if they want it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductPhoto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ProductPhoto'
      security:
      - JWT: []
      summary: Delete Product Photo Information
      tags:
      - ProductPhoto
    put:
      consumes:
      - application/json
      description: User can update product photo information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ProductPhoto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductPhoto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ProductPhoto'
      security:
      - JWT: []
      summary: Update Product Photo Information
      tags:
      - ProductPhoto
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"